+++ exception/src/main/java/hello/exception/exhandler/advice/ExControllerAdvice.java
@@ -1,0 +1,42 @@
+package hello.exception.exhandler.advice;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+import hello.exception.exception.UserException;
+import hello.exception.exhandler.ErrorResult;
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
+@RestControllerAdvice(annotations = RestController.class) // @RestController 애노테이션이 붙어 있는 클래스만 지정
+public class ExControllerAdvice {
+
+	@ResponseStatus(HttpStatus.BAD_REQUEST)
+	@ExceptionHandler(IllegalArgumentException.class)
+	public ErrorResult illegalExHandler(IllegalArgumentException e) {
+		log.error("[exceptionHandler] ex", e);
+
+		return new ErrorResult("BAD", e.getMessage());
+	}
+
+	@ExceptionHandler
+	public ResponseEntity<ErrorResult> userExHandler(UserException e) {
+		log.error("[exceptionHandler] ex", e);
+
+		ErrorResult errorResult = new ErrorResult("USER-EX", e.getMessage());
+		return new ResponseEntity<>(errorResult, HttpStatus.BAD_REQUEST);
+	}
+
+	@ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
+	@ExceptionHandler
+	public ErrorResult exHandler(Exception e) {
+		log.error("[exceptionHandler] ex", e);
+
+		return new ErrorResult("EX", "내부 오류");
+	}
+
+}
